notification_service/requirements.txt
fastapi
uvicorn
kafka-python
jinja2
